+++ frontend/src/Cabinet/assets/data/oAuth.ts
