+++ frontend/src/Cabinet/components/Card/SnsConnectionCard/SnsConnectionCardModal.tsx
@@ -7,7 +7,7 @@ import {
   FailResponseModal,
   SuccessResponseModal,
 } from "@/Cabinet/components/Modals/ResponseModal/ResponseModal";
-import { TOAuthProvider } from "@/Cabinet/assets/data/login";
+import { TOAuthProvider } from "@/Cabinet/assets/data/oAuth";
 import { UserDto } from "@/Cabinet/types/dto/user.dto";
 import IconType from "@/Cabinet/types/enum/icon.type.enum";
 import { axiosMyInfo } from "@/Cabinet/api/axios/axios.custom";
