+++ frontend/src/Cabinet/components/Login/SocialLoginButton.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import LoadingAnimation from "@/Cabinet/components/Common/LoadingAnimation";
 import { ILoginButtonProps } from "@/Cabinet/components/Login/FTLoginButton";
-import { TOAuthProvider } from "@/Cabinet/assets/data/login";
+import { TOAuthProvider } from "@/Cabinet/assets/data/oAuth";
 
 // TODO : FTLoginButton 참고해서 수정하기
