+++ frontend/src/Cabinet/types/dto/login.dto.ts
@@ -1,4 +1,4 @@
-import { TOAuthProvider } from "@/Cabinet/assets/data/login";
+import { TOAuthProvider } from "@/Cabinet/assets/data/oAuth";
 
 // TODO : 주석 수정 필요. 연동 계정 정보 타입 정의
 export interface IUserOAuthConnectionDto {
